Refactor: complete transition to double precision (Phase 1 Priority 4)
- Removed global usage of ImVec2 in domain logic
- Upgraded coordinate systems to glm::dvec2/glm::dvec3 (64-bit precision)
- Implemented drawing bridges in Drawing::Math to cleanly translate double space coordinates to ImVec2 pixel space solely at ImGui boundaries
- Addressed all narrowing and reference type errors across Renderer2D, Canvas, and explicit leaf view implementations
- Removed backup files and updated memory progress tracker

+++ include/shapes/BasicShapes.hpp
@@ -312,8 +312,8 @@ struct Rectangle : public Shape {
              ImU32 color = Colors::RECTANGLE, float thickness = Constants::DEFAULT_LINE_THICKNESS)
         : Shape(ShapeType::RECTANGLE, color, thickness) {
         // Calculate the corners based on the points
-        std::vector<float> xCoords = {p1.x, p2.x, p3.x, p4.x};
-        std::vector<float> yCoords = {p1.y, p2.y, p3.y, p4.y};
+        std::vector<double> xCoords = {p1.x, p2.x, p3.x, p4.x};
+        std::vector<double> yCoords = {p1.y, p2.y, p3.y, p4.y};
         
         auto [minXIt, maxXIt] = std::minmax_element(xCoords.begin(), xCoords.end());
         auto [minYIt, maxYIt] = std::minmax_element(yCoords.begin(), yCoords.end());
